adjust namespaces, ensure RoleManager works, separate public access controller, reduce content controller
