+++ src/Umbraco.Core/Security/IExternalLogin.cs
@@ -1,4 +1,4 @@
-namespace Umbraco.Cms.Core.Models.Identity
+namespace Umbraco.Cms.Core.Security
 {
     /// <summary>
     /// Used to persist external login data for a user
