+++ src/Umbraco.Core/Security/IIdentityUserToken.cs
@@ -1,6 +1,6 @@
 using Umbraco.Cms.Core.Models.Entities;
 
-namespace Umbraco.Cms.Core.Models.Identity
+namespace Umbraco.Cms.Core.Security
 {
     /// <summary>
     /// An external login provider token
