+++ src/Umbraco.Core/Security/IdentityUserToken.cs
@@ -1,7 +1,7 @@
 using System;
 using Umbraco.Cms.Core.Models.Entities;
 
-namespace Umbraco.Cms.Core.Models.Identity
+namespace Umbraco.Cms.Core.Security
 {
     public class IdentityUserToken : EntityBase, IIdentityUserToken
     {
