+++ src/Umbraco.Infrastructure/Persistence/Factories/ExternalLoginFactory.cs
@@ -1,5 +1,5 @@
 using System;
-using Umbraco.Cms.Core.Models.Identity;
+using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Infrastructure.Persistence.Dtos;
 
 namespace Umbraco.Cms.Infrastructure.Persistence.Factories
