+++ src/Umbraco.Infrastructure/Security/IUmbracoUserManager.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
-using Umbraco.Cms.Core.Models.Identity;
 
 namespace Umbraco.Cms.Core.Security
 {
