+++ src/Umbraco.Infrastructure/Services/Implement/ExternalLoginService.cs
@@ -2,9 +2,9 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Extensions.Logging;
 using Umbraco.Cms.Core.Events;
-using Umbraco.Cms.Core.Models.Identity;
 using Umbraco.Cms.Core.Persistence.Repositories;
 using Umbraco.Cms.Core.Scoping;
+using Umbraco.Cms.Core.Security;
 
 namespace Umbraco.Cms.Core.Services.Implement
 {
