+++ src/Umbraco.Web.BackOffice/Mapping/MemberMapDefinition.cs
@@ -6,6 +6,7 @@ using Umbraco.Cms.Core.Mapping;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.ContentEditing;
 using Umbraco.Cms.Core.Models.Mapping;
+using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Web.BackOffice.Trees;
 using Constants = Umbraco.Cms.Core.Constants;
 
@@ -32,6 +33,7 @@ namespace Umbraco.Cms.Web.BackOffice.Mapping
             mapper.Define<IMember, MemberDisplay>((source, context) => new MemberDisplay(), Map);
             mapper.Define<IMember, MemberBasic>((source, context) => new MemberBasic(), Map);
             mapper.Define<IMemberGroup, MemberGroupDisplay>((source, context) => new MemberGroupDisplay(), Map);
+            mapper.Define<UmbracoIdentityRole, MemberGroupDisplay>((source, context) => new MemberGroupDisplay(), Map);
             mapper.Define<IMember, ContentPropertyCollectionDto>((source, context) => new ContentPropertyCollectionDto(), Map);
         }
 
@@ -93,7 +95,17 @@ namespace Umbraco.Cms.Web.BackOffice.Mapping
             target.Path = $"-1,{source.Id}";
             target.Udi = Udi.Create(Constants.UdiEntityType.MemberGroup, source.Key);
         }
-        
+
+        // Umbraco.Code.MapAll -Icon -Trashed -ParentId -Alias -Key -Udi
+        private void Map(UmbracoIdentityRole source, MemberGroupDisplay target, MapperContext context)
+        {
+            target.Id = source.Id;
+            //target.Key = source.Key;
+            target.Name = source.Name;
+            target.Path = $"-1,{source.Id}";
+            //target.Udi = Udi.Create(Constants.UdiEntityType.MemberGroup, source.Key);
+        }
+
         // Umbraco.Code.MapAll
         private static void Map(IMember source, ContentPropertyCollectionDto target, MapperContext context)
         {
