+++ src/Umbraco.Web.BackOffice/Security/IPasswordChanger.cs
@@ -1,7 +1,6 @@
 using System.Threading.Tasks;
 using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Models;
-using Umbraco.Cms.Core.Models.Identity;
 using Umbraco.Cms.Core.Security;
 
 namespace Umbraco.Cms.Web.Common.Security
