+++ src/Umbraco.Web.Common/Security/IMemberRoleManager.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using Umbraco.Cms.Core.Security;
+
+namespace Umbraco.Cms.Web.Common.Security
+{
+    public interface IMemberRoleManager
+    {
+        IEnumerable<UmbracoIdentityRole> Roles { get; }
+    }
+}
